+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - In-App Review test in DebugHomePanel [#2131](https://github.com/rokwire/illinois-app/issues/2131).
 ### Changed
 - Updated "To-Do List" and "Daily Tip" entries from "Wellness" section in Browse panel [#2107](https://github.com/rokwire/illinois-app/issues/2107).
+### Added
+- Fixed audio playback when video tutorial is skipped [#2136](https://github.com/rokwire/illinois-app/issues/2136).
 
 ## [4.1.5] - 2022-07-28
 ### Added