+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [4.1.6] - 2022-07-29
 ### Added
 - In-App Review test in DebugHomePanel [#2131](https://github.com/rokwire/illinois-app/issues/2131).
 ### Changed